COT-47 ajout de CStep::removeAction et ICycle::removeAction Ajout de CAutomate::delAction, mais ne pas laisser l'utilisateur supprimer définitivement une action pour l'instant.

+++ libcot/automate/cycles/CCycleIO.cpp
@@ -27,6 +27,6 @@ void CCycleIO::slotPauseCycle(){
 void CCycleIO::slotStopCycle(){
 }
 
-void CCycleIO::addAction(int, IAction* action){
+void CCycleIO::addAction(float, IAction* action){
 
 }
